Improve calcWorld short→int type for updatedAnimState
Using const short for updatedAnimState caused MWCC to emit an extra
extsh instruction when the lha load already sign-extends the value.
Changing to const int removes the redundant extsh. Score: 96.94% → 97.25%.

+++ src/wm_menu.cpp
@@ -1864,7 +1864,7 @@ void CMenuPcs::calcWorld()
 	model->CalcMatrix();
 	model->CalcSkin();
 
-	const short updatedAnimState = m_wmWorldState->m_mainState;
+	const int updatedAnimState = m_wmWorldState->m_mainState;
 
 	if (updatedAnimState == 1 && animTime >= animEnd) {
 		if (m_wmWorldState->m_frameCounter < 10) {
